+++ client/src/components/pages/information/infoCollection.jsx
@@ -1,0 +1,290 @@
+export default {
+  "My Profile": {
+    title: "My Profile",
+    text: "Everything you need to know about yourself is there.",
+  },
+  "Wanted Hackers": {
+    title: "Wanted Hackers",
+    text: "Everything you need to know about yourself is there.",
+  },
+  "Top Hackers": {
+    title: "Top Hackers",
+    text:
+      "There is a great reward for anyone who shuts down a hacker with a bounty on their head. Be careful, though. Placing a bounty on someone will attract attention to yourself",
+  },
+
+  "Earn Battery": {
+    title: "Earn Battery",
+    text:
+      "A variety of options for you to support CyberhackerWarfare4000 and/or earn some battery",
+  },
+  Information: {
+    title: "Information",
+    text: "You are here right now!",
+  },
+  "Public Forum": {
+    title: "Public Forum",
+    text: "A beta forum is in place for you to communicate to the world!",
+  },
+  Messages: {
+    title: "Messages",
+    text: "Send messages to other hackers.",
+  },
+  Notifications: {
+    title: "Notifications",
+    text:
+      "Whenever you are notified with an event, it will pop up here. Your personal log is also available in this section.",
+  },
+
+  "Top Alliances": {
+    title: "Top Alliances",
+    text: "An overview of the active alliances",
+  },
+  Create: {
+    title: "Create",
+    text: "Form an alliance, if you have the bitcoins for it",
+  },
+  "Alliance Forum": {
+    title: "Alliance Forum",
+    text: "Alliance forum",
+  },
+  Dashboard: {
+    title: "Dashboard",
+    text:
+      "The dashboard shows you and your alliance members aditional information about the alliance you're in. The Boss can do aditional settings in the dashboard",
+  },
+  "Local Hackers": {
+    title: "Local Hackers",
+    text:
+      "An overview of the local hackers in your current city. An icon will indicate if they are online",
+  },
+  "Service & Support": {
+    title: "Service & Support",
+    text:
+      "Heal up your system or hire body guards to protect yourself. Or reset all the statpoints you've used to rebuild your hacker The prices will raise every time you do this.",
+  },
+  VPN: {
+    title: "VPN",
+    text:
+      "Change your VPN status between several cities. Each city have their own prices and datacenters. You can only hack others within the same city",
+  },
+  "Crypto Currency": {
+    title: "Crypto Currency",
+    text:
+      "Arguably the best way to get rich fast. The rules are simple, buy low and sell high. Be careful, though, as your name will show as most recent buyer after purchasing. Prices changes every hour (except at night) and is the same for every city.",
+  },
+  Marketplace: {
+    title: "Marketplace",
+    text:
+      "Boost your system by buying some aditional components. You can only hold one of each category",
+  },
+  Fence: {
+    title: "Fence",
+    text:
+      "Gimmie the loot! The fence is always looking to buy your stolen stash. Head over to the shop and sell whatever you got. The prices varies from each hour and city. Also, the local alliance might take a fee for selling your items in their city!",
+  },
+  Ledger: {
+    title: "Ledger",
+    text:
+      "You can safely store your bitcoins in your ledger. Here it will be safe from being robbed. You can also transfer bitcoins to other users for a small fee",
+  },
+  Petty: {
+    title: "Petty",
+    text:
+      "The easiest and cheapest way to commit crimes. This one is recommended for the new players starting out. You can gain experience, bitcoins and skills by doing petty crimes ",
+  },
+  Crimes: {
+    title: "Crimes",
+    text: "you can commit to earn some bitcoins and xp.",
+  },
+  "Organized Crimes": {
+    title: "Organized Crimes",
+    text:
+      "Commit crimes with your fellow alliance members. Make sure to fill up all the positions in order to get the maximum loot. Upon success, you will be notified with the result.",
+  },
+  Espionage: {
+    title: "Espionage",
+    text:
+      "The espionage is your personal intelligence center. You buy spies and send it to whomever. If the money you spend is more than what the opponent have in their vault, you will be notified with precious information. If it's less however, their vault will be lowered with the amount you spent, and they will be notified that you tried to spy on them.",
+  },
+  "Hack players": {
+    title: "Hack players",
+    text:
+      "Takes you to the local hackers in your city. You can commit attacks on their profile. Make sure to have the right weapon equipped, as it will give you advantage when going into combat!",
+  },
+
+  "Crime skills": {
+    title: "Crime skills",
+    text: (
+      <>
+        There are 4 crime skills:{" "}
+        <span className="text-success"> Technical</span>,
+        <span className="text-success"> Forensics</span>,
+        <span className="text-success"> Social Engineering</span> and
+        <span className="text-success"> Cryptography</span>. These skills can be
+        gained by doing petty crimes, normal crimes or upgrading them with your
+        stat points. The more crime skills you have, the more chance of success
+        you have for commiting crimes and commiting fraud against other hackers!
+      </>
+    ),
+  },
+  /* styling needed */
+  "Hack skills": {
+    title: "Hack skills",
+    text: (
+      <>
+        There are 3 hack skills:<span className="text-primary"> CPU</span>,
+        <span className="text-primary"> AntiVirus</span> and
+        <span className="text-primary"> Encryption</span>. These skills
+        determines your hit/miss chance when going into battle with an opponent
+        hacker.
+        <br />
+        <br />
+        <strong>
+          CPU is effective versus AntiVirus, AntiVirus is effective versus
+          Encryption and Encryption is effective versus CPU.
+        </strong>
+        <br />
+        <br />
+        Make sure to have the right equipped weapon when going into battle - or
+        when suspecting an incoming attack. You can buy components in the market
+        place if you have the bitcoins
+      </>
+    ),
+  },
+  Firewall: {
+    title: "Firewall",
+    text: (
+      <>
+        The <span className="text-danger">firewall</span> stat point represents
+        your overall HP. You start off with 100HP, but this can be upgraded with
+        statpoints or in the market place. Your firewall might go down when
+        attacked by others. But you can repair yourself in the Service &
+        Support.
+      </>
+    ),
+  },
+  Experience: {
+    title: "Experience",
+    text: (
+      <>
+        You gain <span className="text-warning">experience</span> by doing
+        various crimes. The more experience you get, the higher the rank of your
+        hacker will be. With every new rank, you get 5 stat points you can spend
+        in your profile page.
+      </>
+    ),
+  },
+
+  Crimes2: {
+    title: "Crimes",
+    text:
+      "All sorts of crimes will give you both experience and bitcoins (or stash). The higher your crime skills, the more successful you are, which results in more bitcoins and experience.",
+  },
+  Frauds: {
+    title: "Frauds",
+    text: (
+      <>
+        Why earn your own bitcoins when you can steal from others? Use your
+        crime skills and dig deep into the pockets of your enemies! Make sure to
+        protect yourself from other thiefs and secure your bitcoins in your
+        ledger. A fraud will lead to a 5 minute grace period for your target.{" "}
+        <br /> <br /> The success of a fraud is based upon the total of the
+        attackers crime skill.
+      </>
+    ),
+  },
+
+  "Data Centers": {
+    title: "Data Centers",
+    text:
+      "There are 5 data centers in each city throughout the world of cyberhackerwarfare4000. Upon purchase, you will receive a small income every minute. Be sure to keep them healthy by healing them if they get attacked.",
+  },
+
+  Github: {
+    title: "Github",
+    text: (
+      <>
+        <a
+          target="_blank"
+          rel="noopener noreferrer"
+          href="https://github.com/Ojself/cyberwarfare4k"
+        >
+          CyberhackerWarfare4000
+        </a>
+        is an open source project and can be found on Github. You will receive
+        an hourly battery bonus by entering your Github username in the
+        "earn-battery" page and giving the project a star on Github
+      </>
+    ),
+  },
+
+  Patreon: {
+    title: "",
+    text: (
+      <>
+        Support the project at{" "}
+        <a
+          target="_blank"
+          rel="noopener noreferrer"
+          href={
+            "https://www.patreon.com/cyberhackerwarfare4000?fan_landing=true"
+          }
+        >
+          {" "}
+          Patreon
+        </a>{" "}
+        and you will recieve an hourly battery bonus and a huge thank you for
+        supporting open source. The money will go to new features and upkeeping
+        the servers.
+      </>
+    ),
+  },
+
+  Megarpg: {
+    title: "Megarpg",
+    text:
+      "Join the megarpg and type in the code to get a battery bonus. You can do this once every day. The code is case sensitive!",
+  },
+  Chessathor: {
+    title: "Chessathor",
+    text:
+      "Play through the Chessathor game and insert the code (included the #) in the player name input to get a battery bonus. You can do this once every day. The code is case sensitive!",
+  },
+  Attacks: {
+    title: "Attacks",
+    text: (
+      <>
+        If you need to get someone out of the way for good, then attacking is
+        your best option. The equipped weapon of both you and your opponent will
+        determine the hit/miss chance.
+        <br />
+        <br />
+        <strong>Effectiveness</strong>:
+        <br />
+        <span className="text-primary">CPU </span> is effective versus{" "}
+        <span className="text-primary">AntiVirus</span>.{" "}
+        <span className="text-primary">AntiVirus </span>
+        is effective versus <span className="text-primary">Encryption</span>.
+        And <span className="text-primary">Encryption</span> is efective versus{" "}
+        <span className="text-primary">CPU</span>.
+        <br />
+        <br />
+        Damage = (Weapon Skill * effectiveness * 0.1) + (rank level * 1.5)
+        <br />
+        EG: (130 * 1.5 * 0.1) + (7 * 1.5) = <strong>34,5</strong> damage
+        <br />
+        <br /> An attack, successful or not, will grace the player for 60 or 5
+        minutes depending on their online status. If you manage to bring a
+        hacker down to 0 or less firewall, the hacker is shutdown and is forced
+        to start over. You will get whatever crypto currency the player was
+        carrying at the time of a shutdown.
+      </>
+    ),
+  },
+  "Grace period": {
+    title: "Grace period",
+    text:
+      "When someone is attacked, a grace will protect the hacker for a certain time. A fraud will result in a 5 minute grace, while attack will be 5 or 60 minutes grace depending on the online status.",
+  },
+};