+++ server/routes/index.js
@@ -207,17 +207,6 @@ router.get('/opponents/:id', async (req, res) => {
 // Gets all users for the 'Top Hackers' page'
 
 router.get('/ladder', async (req, res) => {
-  /* const dbSelectOptions = {
-    name: '1',
-    alliance: '1',
-    currencies: '1',
-    'playerStats.rankName': '1',
-    'playerStats.rank': '1',
-    'playerStats.bitCoins': '1',
-    'playerStats.ledger': '1',
-    'fightInformation.shutdowns': '1',
-    'fightInformation.crimesInitiated': '1',
-  }; */
   const dbSelectOptions = 'name alliance currencies playerStats.rankName playerStats.rank playerStats.bitCoins playerStats.ledger fightInformation.shutdowns fightInformation.crimesInitiated ';
   let users;
   try {